# Swagger/Cargohub V2/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "CargoHub V2",
      "version": "v2"
    },
    "paths": {
...
      "/api/v2/Shipments": {
        "get": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 1
              }
            },
            {
              "name": "pageSize",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 10
              }
            },
            {
              "name": "sortOrder",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "string",
                "default": "asc"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "post": {
          "tags": [
            "Shipments"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shipment"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shipment"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Shipment"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Shipments/{id}": {
        "get": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "put": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shipment"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shipment"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Shipment"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "patch": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { }
              },
              "text/json": {
                "schema": { }
              },
              "application/*+json": {
                "schema": { }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Shipments/{id}/orders": {
        "get": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "put": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              },
              "application/*+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Shipments/{id}/items": {
        "get": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "put": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              },
              "application/*+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Shipments/{id}/status": {
        "get": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Shipments/search": {
        "get": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "orderId",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "orderDate",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "shipmentStatus",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "carrierCode",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "page",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 1
              }
            },
            {
              "name": "pageSize",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 10
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Shipments/{id}/order": {
        "get": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Shipments/{id}/commit": {
        "put": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Shipments/{id}/force": {
        "delete": {
          "tags": [
            "Shipments"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Item": {
          "type": "object",
          "properties": {
            "Uid": {
              "type": "string",
              "nullable": true
            },
            "Code": {
              "type": "string",
              "nullable": true
            },
            "Description": {
              "type": "string",
              "nullable": true
            },
            "Short_Description": {
              "type": "string",
              "nullable": true
            },
            "Upc_Code": {
              "type": "string",
              "nullable": true
            },
            "Model_Number": {
              "type": "string",
              "nullable": true
            },
            "Commodity_Code": {
              "type": "string",
              "nullable": true
            },
            "Item_Line": {
              "type": "integer",
              "format": "int32"
            },
            "Item_Group": {
              "type": "integer",
              "format": "int32"
            },
            "Item_Type": {
              "type": "integer",
              "format": "int32"
            },
            "Unit_Purchase_Quantity": {
              "type": "integer",
              "format": "int32"
            },
            "Unit_Order_Quantity": {
              "type": "integer",
              "format": "int32"
            },
            "Pack_Order_Quantity": {
              "type": "integer",
              "format": "int32"
            },
            "Supplier_Id": {
              "type": "integer",
              "format": "int32"
            },
            "Supplier_Code": {
              "type": "string",
              "nullable": true
            },
            "Supplier_Part_Number": {
              "type": "string",
              "nullable": true
            },
            "Created_At": {
              "type": "string",
              "nullable": true
            },
            "Updated_At": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
        "Order": {
          "type": "object",
          "properties": {
            "Id": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "Source_Id": {
              "type": "integer",
              "format": "int32"
            },
            "Order_Date": {
              "type": "string",
              "nullable": true
            },
            "Request_Date": {
              "type": "string",
              "nullable": true
            },
            "Reference": {
              "type": "string",
              "nullable": true
            },
            "Reference_Extra": {
              "type": "string",
              "nullable": true
            },
            "Order_Status": {
              "type": "string",
              "nullable": true
            },
            "Notes": {
              "type": "string",
              "nullable": true
            },
            "Shipping_Notes": {
              "type": "string",
              "nullable": true
            },
            "Picking_Notes": {
              "type": "string",
              "nullable": true
            },
            "Warehouse_Id": {
              "type": "integer",
              "format": "int32"
            },
            "Ship_To": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "Bill_To": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "Shipment_Id": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "Total_Amount": {
              "type": "number",
              "format": "double"
            },
            "Total_Discount": {
              "type": "number",
              "format": "double"
            },
            "Total_Tax": {
              "type": "number",
              "format": "double"
            },
            "Total_Surcharge": {
              "type": "number",
              "format": "double"
            },
            "Created_At": {
              "type": "string",
              "nullable": true
            },
            "Updated_At": {
              "type": "string",
              "nullable": true
            },
            "Items": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/OrderItem"
              },
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "OrderItem": {
          "type": "object",
          "properties": {
            "Item_Id": {
              "type": "string",
              "nullable": true
            },
            "Amount": {
              "type": "integer",
              "format": "int32"
            }
          },
          "additionalProperties": false
        },
        "Shipment": {
          "type": "object",
          "properties": {
            "Id": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "Order_Id": {
              "type": "integer",
              "format": "int32"
            },
            "Source_Id": {
              "type": "integer",
              "format": "int32"
            },
            "Order_Date": {
              "type": "string",
              "nullable": true
            },
            "Request_Date": {
              "type": "string",
              "nullable": true
            },
            "Shipment_Date": {
              "type": "string",
              "nullable": true
            },
            "Shipment_Type": {
              "type": "string",
              "nullable": true
            },
            "Shipment_Status": {
              "type": "string",
              "nullable": true
            },
            "Notes": {
              "type": "string",
              "nullable": true
            },
            "Carrier_Code": {
              "type": "string",
              "nullable": true
            },
            "Carrier_Description": {
              "type": "string",
              "nullable": true
            },
            "Service_Code": {
              "type": "string",
              "nullable": true
            },
            "Payment_Type": {
              "type": "string",
              "nullable": true
            },
            "Transfer_Mode": {
              "type": "string",
              "nullable": true
            },
            "Total_Package_Count": {
              "type": "integer",
              "format": "int32"
            },
            "Total_Package_Weight": {
              "type": "number",
              "format": "double"
            },
            "Created_At": {
              "type": "string",
              "nullable": true
            },
            "Updated_At": {
              "type": "string",
              "nullable": true
            },
            "Items": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ShipmentItem"
              },
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "ShipmentItem": {
          "type": "object",
          "properties": {
            "Item_Id": {
              "type": "string",
              "nullable": true
            },
            "Amount": {
              "type": "integer",
              "format": "int32"
            }
          },
          "additionalProperties": false
        },
...
      }
    }
  }